:dizzy: ピンのアイコンを変更した #5
